+++ tests/test_packaging.py
@@ -1,6 +1,6 @@
 def test_version():
     import recast
-    assert recast.__version__ == "0.7.0"
+    assert recast.__version__ == "0.7.1"
 
 
 def test_torchfree_import():